+++ AdapterUnitTests/ChartboostAdapterTests/AUTChartboostRewardedAdTests.m
@@ -13,7 +13,7 @@
 #import "ChartboostSDK.h"
 #endif
 
-#import "GADMAdapterChartboostConstants.h"
+#import "ChartboostAdapter-Swift.h"
 
 typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
@@ -85,9 +85,9 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : location,
-    GADMAdapterChartboostAppID : @"app_id",
-    GADMAdapterChartboostAppSignature : @"signature"
+    [GADMAdapterChartboostConstants adLocation] : location,
+    [GADMAdapterChartboostConstants appID] : @"app_id",
+    [GADMAdapterChartboostConstants appSignature] : @"signature"
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
@@ -128,17 +128,18 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppID : @"app_id",
-    GADMAdapterChartboostAppSignature : @"signature"
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appID] : @"app_id",
+    [GADMAdapterChartboostConstants appSignature] : @"signature"
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
   configuration.credentials = credentials;
 
-  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterChartboostErrorDomain
-                                                      code:201
-                                                  userInfo:nil];
+  NSError *expectedError =
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
+                                 code:201
+                             userInfo:nil];
   AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, configuration, expectedError);
 }
 
@@ -217,9 +218,9 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppID : @"app_id",
-    GADMAdapterChartboostAppSignature : @"signature"
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appID] : @"app_id",
+    [GADMAdapterChartboostConstants appSignature] : @"signature"
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
@@ -232,14 +233,14 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 - (void)testMissingAppID {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppSignature : @"signature",
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appSignature] : @"signature",
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
   configuration.credentials = credentials;
   NSError *expectedError =
-      [[NSError alloc] initWithDomain:GADMAdapterChartboostErrorDomain
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
                                  code:GADMAdapterChartboostErrorInvalidServerParameters
                              userInfo:nil];
   AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, configuration, expectedError);
@@ -248,15 +249,15 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 - (void)testAppIDOnlyWhitespace {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppID : @"    ",
-    GADMAdapterChartboostAppSignature : @"signature",
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appID] : @"    ",
+    [GADMAdapterChartboostConstants appSignature] : @"signature",
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
   configuration.credentials = credentials;
   NSError *expectedError =
-      [[NSError alloc] initWithDomain:GADMAdapterChartboostErrorDomain
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
                                  code:GADMAdapterChartboostErrorInvalidServerParameters
                              userInfo:nil];
   AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, configuration, expectedError);
@@ -265,15 +266,15 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 - (void)testSignatureOnlyWhitespace {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppID : @"app_id",
-    GADMAdapterChartboostAppSignature : @"    ",
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appID] : @"app_id",
+    [GADMAdapterChartboostConstants appSignature] : @"    ",
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
   configuration.credentials = credentials;
   NSError *expectedError =
-      [[NSError alloc] initWithDomain:GADMAdapterChartboostErrorDomain
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
                                  code:GADMAdapterChartboostErrorInvalidServerParameters
                              userInfo:nil];
   AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, configuration, expectedError);
@@ -282,14 +283,14 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 - (void)testMissingSignature {
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppID : @"app_id",
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appID] : @"app_id",
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
   configuration.credentials = credentials;
   NSError *expectedError =
-      [[NSError alloc] initWithDomain:GADMAdapterChartboostErrorDomain
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
                                  code:GADMAdapterChartboostErrorInvalidServerParameters
                              userInfo:nil];
   AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, configuration, expectedError);
@@ -301,15 +302,15 @@ typedef void (^AUTChartboostSetUpCompletionBlock)(CHBStartError *);
 
   AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
   credentials.settings = @{
-    GADMAdapterChartboostAdLocation : @"ad_location",
-    GADMAdapterChartboostAppSignature : @"signature",
-    GADMAdapterChartboostAppID : @"app_id",
+    [GADMAdapterChartboostConstants adLocation] : @"ad_location",
+    [GADMAdapterChartboostConstants appSignature] : @"signature",
+    [GADMAdapterChartboostConstants appID] : @"app_id",
   };
   AUTKMediationRewardedAdConfiguration *configuration =
       [[AUTKMediationRewardedAdConfiguration alloc] init];
   configuration.credentials = credentials;
   NSError *expectedError =
-      [[NSError alloc] initWithDomain:GADMAdapterChartboostErrorDomain
+      [[NSError alloc] initWithDomain:[GADMAdapterChartboostConstants errorDomain]
                                  code:GADMAdapterChartboostErrorMinimumOSVersion
                              userInfo:nil];
   AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, configuration, expectedError);